% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icc.R
\name{icc}
\alias{icc}
\title{ICC (Intraclass Correlation)}
\usage{
icc(x)
}
\arguments{
\item{x}{A model produced using the \code{lme4::lmer()} function. This is an object of class \code{merMod} and subclass \code{lmerMod}.}
}
\value{
A data frame with random effects and their intraclass correlations.
}
\description{
DESCRIPTION

\strong{Note}: For models with random slopes, it is generally advised to interpret with caution. According to Kreft and De Leeuw (2002), "The concept of intra-class correlation is based on a model with a random intercept only. No unique intra-class correlation can be calculated when a random slope is present in the model" (p. 74). However, Snijders and Bosker (YEAR) offer a calculation to derive this value (equation 7.9). This equation is implemented here.
}
\examples{
fit <- lmer(mathach ~ 1 + ses + catholic + (1|id),
data=hsb, REML=T)

icc(fit)

}
\references{
{
\insertRef{hox2018}{mlmhemlpr}
\insertRef{kreft2002}{mlmhemlpr}
}
}
