% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncv_tests.R
\name{ncv_tests}
\alias{ncv_tests}
\title{Non-constant Variance Tests at Level-1 (experimental)}
\usage{
ncv_tests(model, formula = NULL, verbose = FALSE)
}
\arguments{
\item{model}{a mixed model produced using the \code{lme4} package and the \code{lmer()} function. This is an object of class \code{merMod} and subclass \code{lmerMod}. Currently, only supports 2-level models.}

\item{formula}{level-1 formula to compute H test. Formula should be of the form \eqn{y ~ x1 + ... + xn | g} where \eqn{y} is the response, \eqn{x1 + ... + xn} the covariates, and \eqn{g} the grouping factor, see \code{lme4::lmList} for details.}

\item{verbose}{return additional statistics including d-values and outliers from H test; adjusted R^2, ANOVA results, and mean residual by cluster for Levene test; and likelihood ratio test for B-P test.}
}
\value{
A list containing results from the non-constant variance tests.
}
\description{
Computes three different Non-constant variance tests. The H test from Raudenbush and Bryk

An approximate Levene's test discussed by Hox

Finally, a version of the B-P test discussed by X.
}
\examples{

fit <- lme4::lmer(mathach ~ 1 + ses + catholic + (1|id), data=hsb, REML=FALSE)

ncv_tests(fit)

ncv_tests(fit, formula = mathach ~ 1 + ses | id, verbose = TRUE)


}
