% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pve.R
\name{pve}
\alias{pve}
\title{Proportion of variance explained}
\usage{
pve(model1, model2 = NULL)
}
\arguments{
\item{model1}{Previous model, produced using the \code{lme4::lmer()} function. Usually, this is the null or unconditional model.}

\item{model2}{Current model, produced using the \code{lme4::lmer()} function.}
}
\value{
Data frame
}
\description{
\code{pve} calculates the proportional reduction in variance explained (PVE) by adding variables to a prior, nested model. The PVE is considered a local effect size estimate \link{@peugh2020, @raudenbush2002}.
}
\examples{
fit1 <- lme4::lmer(mathach ~ 1 + (1|id), data=hsb, REML=FALSE)
fit2 <- lme4::lmer(mathach ~ 1 + ses + (1|id), data=hsb, REML=FALSE)

pve(fit1, fit2)
}
\references{
{
\insertCite{peugh2010}{mlmhelpr}
\insertCite{raudenbush2002}{mlmhelpr}
}
}
