% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2.R
\name{r2}
\alias{r2}
\title{Pseudo R-squared Measures: squared correlation between predicted and observed values and proportion of variance explained at each level}
\usage{
r2(model1, model2 = NULL)
}
\arguments{
\item{model1}{A model produced using the \code{lme4::lmer()} function. If only model1 is supplied, the squared correlation between predicted and observed values is computed. If both model1 and model2 are supplied, the squared correlation is computed for both models as well as the proportional reduction in variance. In this case, model1 is usually the null or unconditional model.}

\item{model2}{A model produced using the \code{lme4::lmer()} function. Model1 should be nested within model2. Model2 is usually the full or complete model used to compute the proportional reduction in variance based on model1.}
}
\value{
Statement(s) regarding the squared correlation between predicted and observed values and a data frame
}
\description{
The \code{r2} function estimates a pseudo R-squared by correlating predicted \eqn{\hat{Y}} values and observed $Y$ values. This pseudo-$R^2$ is similar to the $R^2$ used in OLS regression. It indicates amount of variation in the outcome that is explained by the model (Peugh, 2010; Singer & Willett, 2003, p. 36). It also calculates the proportional reduction in variance explained (PVE) by adding variables to a prior, nested model. The PVE is considered a local effect size estimate (Peugh, 2010; Raudenbush & Bryk, 2002).
}
\examples{
fit1 <- lme4::lmer(mathach ~ 1 + (1|id), data=hsb, REML=FALSE)
fit2 <- lme4::lmer(mathach ~ 1 + ses + (1|id), data=hsb, REML=FALSE)

r2(fit2)
r2(fit1, fit2)
}
\references{
{
\insertRef{peugh2010}{mlmhelpr}
}

{
\insertRef{raudenbush2002}{mlmhelpr}
}

{
\insertRef{singer2003}{mlmhelpr}
}
}
