% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2_psuedo.R
\name{r2_pseudo}
\alias{r2_pseudo}
\title{Pseudo R-squared: squared correlation between predicted and observed values}
\usage{
r2_pseudo(x, verbose = TRUE)
}
\arguments{
\item{x}{A model produced using the \code{lme4::lmer()} function. This is an object of class \code{merMod} and subclass \code{lmerMod}.}

\item{verbose}{If true (defualt), prints an explanatory message, "The squared correlation between predicted and observed values is...". If false, returns a value.}
}
\value{
If \code{verbose == T} (default), a console message. If \code{verbose == F}, a numeric value.
}
\description{
The \code{r2} function estimates a pseudo R-squared by correlating predicted \eqn{\hat{Y}} values and observed $Y$ values. This pseudo-$R^2$ is similar to the $R^2$ used in OLS regression. It indicates amount of variation in the outcome that is explained by the model \link{@peugh2020; @singer2003, p. 36}.
}
\examples{

fit <- lme4::lmer(mathach ~ 1 + ses + catholic + (1|id),
data=hsb, REML=TRUE)

# returns a console message with the r2 value
r2_cor(fit)

# returns a numeric value
r2_cor(fit, verbose = FALSE)
r2_cor(fit, FALSE)
}
\references{
{
\insertCite{peugh2010}{mlmhelpr}
\insertCite{singer2003}{mlmhelpr}
}
}
